+++ packages/gatsby-plugin-fastify/src/plugins/500.ts
@@ -9,7 +9,7 @@ export const handle500: FastifyPluginAsync = fp(async (fastify, _opts) => {
   fastify.log.info(
     `Gatsby 500 error page ${
       gatsby500ErrorFileExists ? "exists" : "missing, using generic 500 error for DSG/SSR"
-    }`,
+    }`
   );
 
   fastify.setErrorHandler(async (err, req, reply) => {