+++ packages/gatsby-plugin-fastify/src/plugins/serverRoutes.ts
@@ -23,7 +23,7 @@ export const handleServerRoutes: FastifyPluginAsync<{
         }
         return acc;
       },
-      { dsgCount: 0, ssrCount: 0 },
+      { dsgCount: 0, ssrCount: 0 }
     );
 
     fastify.log.info(`Registering ${dsgCount} DSG route(s)`);
@@ -67,11 +67,16 @@ export const handleServerRoutes: FastifyPluginAsync<{
             graphqlEngine,
             req,
           });
+
           const pageData = (await renderPageData({ data: pageQueryData })) as any;
 
-          if (page.mode === `SSR` && pageQueryData?.serverDataHeaders) {
-            for (const [name, value] of Object.entries(pageQueryData.serverDataHeaders)) {
-              reply.header(name, value);
+          if (page.mode === `SSR`) {
+            if (pageQueryData?.serverDataHeaders) {
+              reply.headers(pageQueryData.serverDataHeaders);
+            }
+
+            if (pageQueryData?.serverDataStatus) {
+              reply.code(pageQueryData.serverDataStatus);
             }
           }
 
@@ -102,15 +107,21 @@ export const handleServerRoutes: FastifyPluginAsync<{
           appendModuleHeader(page?.mode as "DSG" | "SSR", reply);
 
           try {
-            const data = await getData({
+            const pageQueryData = await getData({
               pathName: potentialPagePath,
               graphqlEngine,
               req,
             });
-            const results = await renderHTML({ data });
-            if (page.mode === `SSR` && data.serverDataHeaders) {
-              for (const [name, value] of Object.entries(data.serverDataHeaders)) {
-                reply.header(name, value);
+
+            const results = await renderHTML({ data: pageQueryData });
+
+            if (page.mode === `SSR`) {
+              if (pageQueryData?.serverDataHeaders) {
+                reply.headers(pageQueryData.serverDataHeaders);
+              }
+
+              if (pageQueryData?.serverDataStatus) {
+                reply.code(pageQueryData.serverDataStatus);
               }
             }