+++ packages/gatsby-plugin-github-ribbon/jest.config.js
@@ -15,4 +15,4 @@ module.exports = {
 
   // The test environment that will be used for testing
   testEnvironment: `node`,
-}
+};