+++ test-sites/fastify/src/pages/index.js
@@ -1,46 +1,46 @@
-import * as React from "react"
-import { Link, withPrefix } from "gatsby"
+import * as React from "react";
+import { Link, withPrefix } from "gatsby";
 // styles
 const pageStyles = {
   color: "#232129",
   padding: 96,
   fontFamily: "-apple-system, Roboto, sans-serif, serif",
-}
+};
 const headingStyles = {
   marginTop: 0,
   marginBottom: 64,
   maxWidth: 320,
-}
+};
 const headingAccentStyles = {
   color: "#663399",
-}
+};
 const paragraphStyles = {
   marginBottom: 48,
-}
+};
 const codeStyles = {
   color: "#8A6534",
   padding: 4,
   backgroundColor: "#FFF4DB",
   fontSize: "1.25rem",
   borderRadius: 4,
-}
+};
 const listStyles = {
   marginBottom: 96,
   paddingLeft: 0,
-}
+};
 const listItemStyles = {
   fontWeight: 300,
   fontSize: 24,
   maxWidth: 560,
   marginBottom: 30,
-}
+};
 
 const linkStyle = {
   color: "#8954A8",
   fontWeight: "bold",
   fontSize: 16,
   verticalAlign: "5%",
-}
+};
 
 // const descriptionStyle = {
 //   color: "#232129",
@@ -65,7 +65,7 @@ const IndexPage = () => {
       title: "Page 3",
       uri: "/posts/page-3",
     },
-  ]
+  ];
 
   return (
     <main style={pageStyles}>
@@ -79,8 +79,8 @@ const IndexPage = () => {
         </span>
       </h1>
       <p style={paragraphStyles}>
-        Edit <code style={codeStyles}>src/pages/index.js</code> to see this page
-        update in real-time.{" "}
+        Edit <code style={codeStyles}>src/pages/index.js</code> to see this page update in
+        real-time.{" "}
         <span role="img" aria-label="Sunglasses smiley emoji">
           😎
         </span>
@@ -131,9 +131,12 @@ const IndexPage = () => {
         <li style={{ ...listItemStyles }}>
           <a href={withPrefix("/ssrBad")}>Bad SSR Page</a>
         </li>
+        <li style={{ ...listItemStyles }}>
+          <a href={withPrefix("/ssr403")}>Unauthorized SSR Page</a>
+        </li>
       </ul>
     </main>
-  )
-}
+  );
+};
 
-export default IndexPage
+export default IndexPage;