+++ test-sites/fastify/src/templates/example.js
@@ -1,5 +1,5 @@
-import * as React from "react"
+import * as React from "react";
 
 export default function Example({ pageContext }) {
-  return <div>Hello world #{pageContext.pageNumber}!</div>
+  return <div>Hello world #{pageContext.pageNumber}!</div>;
 }